refactor: frame 추상클래스로 변경
