+++ src/main/java/bowling/domain/frame/InitialFrame.java
@@ -1,25 +1,17 @@
 package bowling.domain.frame;
 
-import bowling.domain.dto.StateData;
 import bowling.domain.score.Score;
 import bowling.domain.state.CommonState;
 import bowling.domain.state.pitching.FirstPitching;
 
-public class InitialFrame implements Frame {
-
-    private final CommonState state;
+public class InitialFrame extends Frame {
 
     private InitialFrame(CommonState state) {
-        this.state = state;
+        super(state);
     }
 
     public static InitialFrame of() {
         return new InitialFrame(FirstPitching.of());
-    }
-
-    @Override
-    public boolean isBowlingFinish() {
-        return false;
     }
 
     @Override
@@ -28,12 +20,8 @@ public class InitialFrame implements Frame {
     }
 
     @Override
-    public Score addBonusScore(Score score) {
+    protected Score addBonusScore(Score score) {
         return score;
     }
 
-    @Override
-    public StateData getFrameStates() {
-        return null;
-    }
 }
